convert to BigDecimal

+++ src/main/java/com/flowyk/neuron/messenger/ActivationResult.java
@@ -1,16 +1,18 @@
 package com.flowyk.neuron.messenger;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Lukas on 30. 11. 2014.
  */
 public class ActivationResult {
 
-    private double output;
-    public ActivationResult(double output) {
+    private BigDecimal output;
+    public ActivationResult(BigDecimal output) {
         this.output = output;
     }
 
-    public double getOutput() {
+    public BigDecimal getOutput() {
         return output;
     }
 }
